Docs

+++ Assets/GradientOverlay.shader
@@ -14,16 +14,16 @@
     sampler2D _MainTex;
 
     float4x4 _Gradient;
-    half _Opacity;
-    half2 _Direction;
+    float _Opacity;
+    float2 _Direction;
     float _Frequency;
     float _Scroll;
     float _NoiseStrength;
     float _NoiseAnimation;
 
-    half4 frag(v2f_img i) : SV_Target
+    float4 frag(v2f_img i) : SV_Target
     {
-        // Base animation.
+        // Base animation
         float2 uv = i.uv - 0.5;
         uv.x *= _ScreenParams.x * (_ScreenParams.w - 1);
         uv += _Direction * _Scroll;
@@ -31,17 +31,18 @@
 
         float p = dot(uv, _Direction);
 
-        // Noise field.
+        // Noise field
         p = lerp(p, snoise(float3(uv, _NoiseAnimation)), _NoiseStrength);
 
         // Pick color from the gradient.
-        half3 rgb = CosineGradient(_Gradient, p);
+        float3 rgb = CosineGradient(_Gradient, p);
+
         #if !defined(UNITY_COLORSPACE_GAMMA)
         rgb = GammaToLinearSpace(rgb);
         #endif
 
         // Mix with the source.
-        half4 col = tex2D(_MainTex, i.uv);
+        float4 col = tex2D(_MainTex, i.uv);
         col.rgb = lerp(col.rgb, rgb, _Opacity);
         return col;
     }
